+++ src/Usenet/Nntp/Parsers/GroupResponseParser.cs
@@ -20,7 +20,7 @@ namespace Usenet.Nntp.Parsers
                 return new NntpGroupResponse(
                     code, message, false,
                     new NntpGroup(string.Empty, 0, 0, 0, NntpPostingStatus.Unknown,
-                        string.Empty, new int[0]));
+                        string.Empty, new long[0]));
             }
 
             string[] responseSplit = message.Split(' ');
@@ -29,15 +29,15 @@ namespace Usenet.Nntp.Parsers
                 log.LogError("Invalid response message: {Message} Expected: {{count}} {{low}} {{high}} {{group}}", message);
             }
 
-            int.TryParse(responseSplit.Length > 0 ? responseSplit[0] : null, out int articleCount);
-            int.TryParse(responseSplit.Length > 1 ? responseSplit[1] : null, out int lowWaterMark);
-            int.TryParse(responseSplit.Length > 2 ? responseSplit[2] : null, out int highWaterMark);
+            long.TryParse(responseSplit.Length > 0 ? responseSplit[0] : null, out long articleCount);
+            long.TryParse(responseSplit.Length > 1 ? responseSplit[1] : null, out long lowWaterMark);
+            long.TryParse(responseSplit.Length > 2 ? responseSplit[2] : null, out long highWaterMark);
             string name = responseSplit.Length > 3 ? responseSplit[3] : string.Empty;
 
             return new NntpGroupResponse(
                 code, message, true,
                 new NntpGroup(name, articleCount, lowWaterMark, highWaterMark, NntpPostingStatus.Unknown,
-                    string.Empty, new int[0]));
+                    string.Empty, new long[0]));
         }
     }
 }
