+++ src/Usenet/Nntp/Parsers/NextResponseParser.cs
@@ -36,7 +36,7 @@ namespace Usenet.Nntp.Parsers
                 log.LogError("Invalid response message: {Message} Expected: {{number}} {{messageid}}", message);
             }
 
-            int.TryParse(responseSplit.Length > 0 ? responseSplit[0] : null, out int number);
+            long.TryParse(responseSplit.Length > 0 ? responseSplit[0] : null, out long number);
             string messageId = responseSplit.Length > 1 ? responseSplit[1] : string.Empty;
 
             return new NntpNextResponse(code, message, true, responseType, number, messageId);
