+++ src/Usenet/Nntp/Responses/NntpLastResponse.cs
@@ -17,7 +17,7 @@ namespace Usenet.Nntp.Responses
         /// The highest existing <see cref="NntpArticle"/> number less than the 
         /// current <see cref="NntpArticle"/> in the currently selected <see cref="NntpGroup"/>.
         /// </summary>
-        public int Number { get; }
+        public long Number { get; }
 
         /// <summary>
         /// The <see cref="NntpMessageId"/> of the <see cref="NntpArticle"/> with the highest existing number less than the 
@@ -34,7 +34,7 @@ namespace Usenet.Nntp.Responses
         /// <param name="responseType">The type of the response received from the server.</param>
         /// <param name="number">The <see cref="NntpArticle"/> number received from the server.</param>
         /// <param name="messageId">The <see cref="NntpMessageId"/> received from the server.</param>
-        public NntpLastResponse(int code, string message, bool success, NntpLastResponseType responseType, int number, NntpMessageId messageId)
+        public NntpLastResponse(int code, string message, bool success, NntpLastResponseType responseType, long number, NntpMessageId messageId)
             : base (code, message, success)
         {
             ResponseType = responseType;
