+++ src/Usenet/Nntp/Responses/NntpStatResponse.cs
@@ -16,7 +16,7 @@ namespace Usenet.Nntp.Responses
         /// <summary>
         /// The <see cref="NntpArticle"/> number received from the server.
         /// </summary>
-        public int Number { get; }
+        public long Number { get; }
 
         /// <summary>
         /// The <see cref="NntpMessageId"/> received from the server.
@@ -32,7 +32,7 @@ namespace Usenet.Nntp.Responses
         /// <param name="responseType">The type of the response received from the server.</param>
         /// <param name="number">The <see cref="NntpArticle"/> number received from the server.</param>
         /// <param name="messageId">The <see cref="NntpMessageId"/> received from the server.</param>
-        public NntpStatResponse(int code, string message, bool success, NntpStatResponseType responseType, int number, NntpMessageId messageId)
+        public NntpStatResponse(int code, string message, bool success, NntpStatResponseType responseType, long number, NntpMessageId messageId)
             : base (code, message, success)
         {
             ResponseType = responseType;
