+++ test/UsenetTests/Nntp/Parsers/GroupsResponseParserTests.cs
@@ -22,8 +22,8 @@ namespace UsenetTests.Nntp.Parsers
                 },
                 new XSerializable<NntpGroup[]>(new []
                 {
-                    new NntpGroup("alt.rfc-writers.recovery", 0, 1, 4, NntpPostingStatus.PostingPermitted, string.Empty, new List<int>(0)),
-                    new NntpGroup("tx.natives.recovery", 0, 56, 89, NntpPostingStatus.PostingPermitted, string.Empty, new List<int>(0))
+                    new NntpGroup("alt.rfc-writers.recovery", 0, 1, 4, NntpPostingStatus.PostingPermitted, string.Empty, new List<long>(0)),
+                    new NntpGroup("tx.natives.recovery", 0, 56, 89, NntpPostingStatus.PostingPermitted, string.Empty, new List<long>(0))
                 })
             },
             new object[]
@@ -41,8 +41,8 @@ namespace UsenetTests.Nntp.Parsers
                 },
                 new XSerializable<NntpGroup[]>(new []
                 {
-                    new NntpGroup("misc.test", 1234, 3000234, 3002322, NntpPostingStatus.PostingPermitted, string.Empty, new List<int>(0)),
-                    new NntpGroup("rec.food.drink.tea", 3, 51, 100, NntpPostingStatus.PostingPermitted, string.Empty, new List<int>(0))
+                    new NntpGroup("misc.test", 1234, 3000234, 3002322, NntpPostingStatus.PostingPermitted, string.Empty, new List<long>(0)),
+                    new NntpGroup("rec.food.drink.tea", 3, 51, 100, NntpPostingStatus.PostingPermitted, string.Empty, new List<long>(0))
                 })
             },
             new object[]
